Remove wildcards, pull out LastElementBottom arrangement
