+++ feature/tabsflow/moreflow/faq/src/main/java/com/epmedu/animeal/tabs/more/faq/presentation/FAQScreenUI.kt
@@ -1,23 +1,25 @@
 package com.epmedu.animeal.tabs.more.faq.presentation
 
-import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.statusBarsPadding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
-import androidx.compose.material.*
+import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.unit.Density
 import androidx.compose.ui.unit.dp
+import com.epmedu.animeal.foundation.layout.LastElementBottom
 import com.epmedu.animeal.foundation.preview.AnimealPreview
 import com.epmedu.animeal.foundation.theme.AnimealTheme
 import com.epmedu.animeal.resources.R
+import com.epmedu.animeal.tabs.more.faq.domain.model.FAQCard
 import com.epmedu.animeal.tabs.more.faq.presentation.FAQScreenEvent.BackClicked
 import com.epmedu.animeal.tabs.more.faq.presentation.FAQScreenEvent.CardClicked
-import com.epmedu.animeal.tabs.more.faq.domain.model.FAQCard
 import com.epmedu.animeal.tabs.more.faq.presentation.ui.FAQHeader
 import com.epmedu.animeal.tabs.more.faq.presentation.ui.FAQListItem
 import com.epmedu.animeal.tabs.more.faq.presentation.viewmodel.FAQState
@@ -34,26 +36,7 @@ internal fun FAQScreenUI(
             .fillMaxSize()
             .statusBarsPadding(),
         horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = remember {
-            // Sticks last element to the bottom
-            object : Arrangement.Vertical {
-                override fun Density.arrange(
-                    totalSize: Int,
-                    sizes: IntArray,
-                    outPositions: IntArray
-                ) {
-                    var currentOffset = 0
-                    sizes.forEachIndexed { index, size ->
-                        if (index == sizes.lastIndex) {
-                            outPositions[index] = totalSize - size
-                        } else {
-                            outPositions[index] = currentOffset
-                            currentOffset += size
-                        }
-                    }
-                }
-            }
-        },
+        verticalArrangement = Arrangement.LastElementBottom,
     ) {
         item { FAQHeader(onBack = { onEvent(BackClicked) }) }
         items(state.questionCards) {
